Update class-learnpress-discord-addon-admin.php
Separate updating the default role to handle the case where the student has no enrolled courses AND the default role is defined.
Add test to catch the empty role id.

+++ admin/class-learnpress-discord-addon-admin.php
@@ -530,57 +530,66 @@ class Learnpress_Discord_Addon_Admin {
 		if ( $access_token && $refresh_token && is_array( $ets_learnpress_discord_role_mapping ) && is_array( $student_courses ) ){
 			foreach ( $student_courses as $course_id ) {                    
 		
-			$student_role_for_course  = get_user_meta( $user_id,'_ets_learnpress_discord_role_id_for_' . $course_id , true);
-                        
-			if( $student_role_for_course && array_key_exists( 'learnpress_course_id_' . $course_id, $ets_learnpress_discord_role_mapping ) ){
-                            
-				// Nothing to do;
-                    
+                            $student_role_for_course  = get_user_meta( $user_id,'_ets_learnpress_discord_role_id_for_' . $course_id , true);
+
+                            if( $student_role_for_course && array_key_exists( 'learnpress_course_id_' . $course_id, $ets_learnpress_discord_role_mapping ) ){
+
+                                    // Nothing to do;
+
+                            }
+                            if( $student_role_for_course && array_key_exists( 'learnpress_course_id_' . $course_id, $ets_learnpress_discord_role_mapping ) && $ets_learnpress_discord_role_mapping['learnpress_course_id_' . $course_id] != $student_role_for_course ){
+
+                                    // Remove $student_role_for_course
+                                    $old_role = $student_role_for_course;
+                                    delete_user_meta( $user_id, '_ets_learnpress_discord_role_id_for_' . $course_id , $old_role );
+                                    $this->learnpress_discord_public_instance->delete_discord_role( $user_id, $old_role );
+                                    update_option('runapi_1_'.$course_id.'_'.$old_role.'_'.rand(), $old_role);
+
+                                    // Assign $ets_learnpress_discord_role_mapping['learnpress_course_id_' . $course_id]
+                                    $new_role = $ets_learnpress_discord_role_mapping['learnpress_course_id_' . $course_id];
+                                    update_user_meta( $user_id, '_ets_learnpress_discord_role_id_for_' . $course_id , $new_role );
+                                    $this->learnpress_discord_public_instance->put_discord_role_api( $user_id, $new_role );
+                                    update_option('runapi_2_'.$course_id.'_'.$new_role.'_'.rand(), $new_role);
+
+                            }                        
+
+                            if( ! $student_role_for_course && array_key_exists( 'learnpress_course_id_' . $course_id, $ets_learnpress_discord_role_mapping ) ){
+
+                                    $new_role = $ets_learnpress_discord_role_mapping['learnpress_course_id_' . $course_id];
+                                    update_user_meta( $user_id, '_ets_learnpress_discord_role_id_for_' . $course_id , $new_role );
+                                    $this->learnpress_discord_public_instance->put_discord_role_api( $user_id, $new_role );
+                                    update_option('runapi_3_'.$course_id.'_'.$new_role.'_'.rand(), $new_role);
+                            }
+
+                            if ( $student_role_for_course && ! array_key_exists( 'learnpress_course_id_' . $course_id, $ets_learnpress_discord_role_mapping ) ){
+
+                                    $old_role = $student_role_for_course;
+                                    delete_user_meta( $user_id, '_ets_learnpress_discord_role_id_for_' . $course_id , $old_role ); 
+                                    $this->learnpress_discord_public_instance->delete_discord_role( $user_id, $old_role ); 
+                                    update_option('runapi_4_'.$course_id.'_'.$old_role.'_'.rand(), $old_role);
+                            }
 			}
-			if( $student_role_for_course && array_key_exists( 'learnpress_course_id_' . $course_id, $ets_learnpress_discord_role_mapping ) && $ets_learnpress_discord_role_mapping['learnpress_course_id_' . $course_id] != $student_role_for_course ){
 
-				// Remove $student_role_for_course
-				$old_role = $student_role_for_course;
-				delete_user_meta( $user_id, '_ets_learnpress_discord_role_id_for_' . $course_id , $old_role );
-				$this->learnpress_discord_public_instance->delete_discord_role( $user_id, $old_role );
-                            
-				// Assign $ets_learnpress_discord_role_mapping['learnpress_course_id_' . $course_id]
-				$new_role = $ets_learnpress_discord_role_mapping['learnpress_course_id_' . $course_id];
-				update_user_meta( $user_id, '_ets_learnpress_discord_role_id_for_' . $course_id , $new_role );
-				$this->learnpress_discord_public_instance->put_discord_role_api( $user_id, $new_role ); 
-                    
-                        }                        
+		}
+ 		if ( $access_token && $refresh_token ){
+                    if ( $default_role && $default_role != 'none' && $default_role === $last_default_role ){
+                        //
 
-			if( ! $student_role_for_course && array_key_exists( 'learnpress_course_id_' . $course_id, $ets_learnpress_discord_role_mapping ) ){
-			
-				$new_role = $ets_learnpress_discord_role_mapping['learnpress_course_id_' . $course_id];
-				update_user_meta( $user_id, '_ets_learnpress_discord_role_id_for_' . $course_id , $new_role );
-				$this->learnpress_discord_public_instance->put_discord_role_api( $user_id, $new_role );                             
-			}
+                    }elseif ( $default_role && $default_role != 'none' && $default_role != $last_default_role  ) {
 
-			if ( $student_role_for_course && ! array_key_exists( 'learnpress_course_id_' . $course_id, $ets_learnpress_discord_role_mapping ) ){
-                            
-				$old_role = $student_role_for_course;
-				delete_user_meta( $user_id, '_ets_learnpress_discord_role_id_for_' . $course_id , $old_role ); 
-				$this->learnpress_discord_public_instance->delete_discord_role( $user_id, $old_role );                            
-			}
-                }
+                            update_user_meta( $user_id, '_ets_learnpress_discord_last_default_role', $default_role );
+                            $this->learnpress_discord_public_instance->delete_discord_role( $user_id, $last_default_role );
+                            $this->learnpress_discord_public_instance->put_discord_role_api( $user_id, $default_role );
+                            update_option('runapi_5', $last_default_role);
+                            update_option('runapi_6', $default_role);
+                    }else{
 
-
-		if ( $default_role && $default_role != 'none' && $default_role === $last_default_role ){
-                    //
-                    
-		}elseif ( $default_role && $default_role != 'none' && $default_role != $last_default_role  ) {
-                    
-			update_user_meta( $user_id, '_ets_learnpress_discord_last_default_role', $default_role );
-			$this->learnpress_discord_public_instance->delete_discord_role( $user_id, $last_default_role );
-			$this->learnpress_discord_public_instance->put_discord_role_api( $user_id, $default_role );
-		}else{
-			
-			delete_user_meta( $user_id, '_ets_learnpress_discord_last_default_role' );
-			$this->learnpress_discord_public_instance->delete_discord_role( $user_id, $last_default_role );   
-		}                
-	}        
+                            delete_user_meta( $user_id, '_ets_learnpress_discord_last_default_role' );
+                            $this->learnpress_discord_public_instance->delete_discord_role( $user_id, $last_default_role );
+                            update_option('runapi_7', $last_default_role);
+                    }                    
+ 		}                       
+        
 	exit();
            
         }
